Get localize config schema working with modules
The schema generator doesn't work well with modules, but by extracting the
pure types into their own directory, we can have an isolated build that it's
happy with. This has the side effect of being much much faster too.
