+++ packages/localize-tools/src/cli.ts
@@ -7,14 +7,13 @@
 import * as path from 'path';
 import minimist from 'minimist';
 import {KnownError, unreachable} from './error.js';
-import {Config, readConfigFileAndWriteSchema} from './config.js';
+import {Config} from './types/config.js';
+import {readConfigFileAndWriteSchema} from './config.js';
 import {LitLocalizer} from './index.js';
 import {printDiagnostics} from './typescript.js';
-import {
-  TransformLitLocalizer,
-  TransformOutputConfig,
-} from './modes/transform.js';
-import {RuntimeLitLocalizer, RuntimeOutputConfig} from './modes/runtime.js';
+import {TransformLitLocalizer} from './modes/transform.js';
+import {RuntimeLitLocalizer} from './modes/runtime.js';
+import {TransformOutputConfig, RuntimeOutputConfig} from './types/modes.js';
 import {dirname} from 'path';
 import {fileURLToPath} from 'url';
 import * as fs from 'fs/promises';
