+++ packages/localize-tools/src/types/locale.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @license
+ * Copyright 2021 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+/** Opaque type for locale IDs */
+export type Locale = string & {__TYPE__: 'Locale'};
