+++ packages/localize-tools/tsconfig.schema.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
-    "target": "es2018",
-    "moduleResolution": "node",
-    "lib": ["es2018", "DOM"]
+    "target": "es2020",
+    "lib": ["es2020"],
+    "types": [],
   },
-  "include": ["src/**/*"]
+  "include": ["src/types/**/*"],
+  "exclude": ["node_modules"],
 }
